requirements.txt
torch==2.3.1
torchvision==0.18.1
numpy>=1.26
scipy>=1.11
pyyaml>=6.0
tqdm>=4.66
